Add a portable replacement for getgrouplist()
Not all platforms include an interface like getgrouplist() or
getgroupsbymember(). On those platforms, let's fall back to the crudest
possible way of enumerating all groups and selecting those that have
the member we are interested in.

Note that since this interface requires the getgrent() interface to be
available which might not even be possible if the LDAP client is
configured against the slapi-NIS server of FreeIPA.